complete logout and users folders

+++ index.php
@@ -2,6 +2,19 @@
 
 include "bootstrap/init.php";
 
+if (isset($_GET['logout'])) {
+    // redirect to aut form
+    logout();
+}
+
+if (!isLoggedIn()) {
+    // redirect to aut form
+    
+    redirect(site_url('auth.php'));
+}
+
+$user = getLoggedInUser();
+
 if(isset($_GET['delete_folder']) && is_numeric($_GET['delete_folder'])){
     $deletedCount = deleteFolder($_GET['delete_folder']);
     // echo "$deletedCount folders succesfully deleted!";
